requirements.txt
imageio==2.6.1
matplotlib==3.1.1
numpy==1.12.1
opencv_python==3.4.0.12
pandas==1.1.5
Pillow==8.4.0
PyYAML==6.0
scikit_image==0.15.0
scipy==1.2.1
timm==0.4.12
torch==1.1.0
torchvision==0.3.0
tqdm==4.31.1
